requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
